feat: 完善库存精度与工作台风险预览

+++ erp-server/erp-product/src/main/java/com/qiheng/erp/product/domain/entity/Product.java
@@ -87,7 +87,7 @@ public class Product implements Serializable {
     @Schema(description = "安全库存数量，按100倍整数存储，例如12.50存1250")
     @TableField("safety_stock_qty")
     @NotNull(message = "安全库存数量不能为空")
-    private BigDecimal safetyStockQty;
+    private Long safetyStockQty;
 
     @Schema(description = "状态：1启用，0禁用")
     @TableField("status")
